move the com.example.goodfood to com.hmir.goodfood which can stand for our app better
