+++ app/src/androidTest/java/com/hmir/goodfood/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.goodfood;
+package com.hmir.goodfood;
 
 import android.content.Context;
