+++ app/src/main/java/com/hmir/goodfood/Calories.java
@@ -1,4 +1,4 @@
-package com.example.goodfood;
+package com.hmir.goodfood;
 
 import android.os.Bundle;
 import android.widget.TextView;
