+++ app/src/main/java/com/hmir/goodfood/ExtractIngredient.java
@@ -1,4 +1,4 @@
-package com.example.goodfood;
+package com.hmir.goodfood;
 
 import android.content.Intent;
 import android.os.Bundle;
