+++ app/src/main/java/com/hmir/goodfood/Meal.java
@@ -1,4 +1,4 @@
-package com.example.goodfood;
+package com.hmir.goodfood;
 
 public class Meal {
     // Temporary meal class for data
