+++ app/src/main/java/com/hmir/goodfood/RoundedBarChart.java
@@ -1,16 +1,13 @@
-package com.example.goodfood;
+package com.hmir.goodfood;
 
 import android.graphics.Canvas;
-import android.graphics.LinearGradient;
 import android.graphics.Path;
 import android.graphics.RectF;
-import android.graphics.Shader;
 
 import com.github.mikephil.charting.animation.ChartAnimator;
 import com.github.mikephil.charting.buffer.BarBuffer;
 import com.github.mikephil.charting.interfaces.dataprovider.BarDataProvider;
 import com.github.mikephil.charting.interfaces.datasets.IBarDataSet;
-import com.github.mikephil.charting.model.GradientColor;
 import com.github.mikephil.charting.renderer.BarChartRenderer;
 import com.github.mikephil.charting.utils.Transformer;
 import com.github.mikephil.charting.utils.ViewPortHandler;
