+++ app/src/main/res/layout/activity_dashboard.xml
@@ -53,7 +53,7 @@
 
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/fragmentContainerView3"
-            android:name="com.example.goodfood.BottomNavigationFragment"
+            android:name="com.hmir.goodfood.BottomNavigationFragment"
             android:layout_width="match_parent"
             android:layout_height="match_parent" />
     </FrameLayout>
