+++ app/src/main/res/layout/activity_extract_ingredient.xml
@@ -80,7 +80,7 @@
 
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/fragmentContainerView"
-            android:name="com.example.goodfood.BottomNavigationFragment"
+            android:name="com.hmir.goodfood.BottomNavigationFragment"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content" />
     </FrameLayout>
