+++ app/src/main/res/layout/activity_food_scanner.xml
@@ -94,7 +94,7 @@
 
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/fragmentContainerView"
-            android:name="com.example.goodfood.BottomNavigationFragment"
+            android:name="com.hmir.goodfood.BottomNavigationFragment"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content" />
     </FrameLayout>
