+++ app/src/test/java/com/hmir/goodfood/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.goodfood;
+package com.hmir.goodfood;
 
 import org.junit.Test;
